pyproject.toml
[project]
name = "splitwise-mcp"
version = "0.1.0"
description = "MCP server for Splitwise — manage expenses, groups, and friends from any LLM client"
readme = "README.md"
authors = [
    { name = "Rohan", email = "rohan26gupta2005@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
splitwise-mcp = "splitwise_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
